+++ edgy_guardian/content_types/utils.py
@@ -21,4 +21,4 @@ async def get_content_type(obj: Any) -> type[edgy.Model]:
     ContentType
         The content type of the object.
     """
-    return await get_content_type_model().query.get_for_model(obj)
+    return await get_content_type_model().guardian.get_for_model(obj)