+++ docs/release-notes.md
@@ -2,12 +2,22 @@
 
 ## 0.1.1
 
+### Added
+
+- [Managers](./managers.md) documentation section.
+- Allow passing objects for permissions and groups in `has_user_perm` and `has_group_permission`.
+
+### Changed
+
+- Rename internal models manager default from `query` to `guardian`. This will avoid any clash or
+the existing `query` override in any edgy models.
+
 ### Fixed
 
 - Unnecessary load of apps for ContentTypes.
 - Mandatory obj attribute in `assign_perm`.
 - Mypy typing and checks
-  
+
 ## 0.1.0
 
 ### Added